+++ test/ex_oauth2_provider/redirect_uri_test.exs
@@ -49,19 +49,28 @@ defmodule ExOauth2Provider.RedirectURITest do
 
   test "matches?#true" do
     uri = "https://app.co/aaa"
-    assert RedirectURI.matches?(uri, uri)
+    assert RedirectURI.matches?(uri, uri, [])
+  end
+
+  test "matches?#true with custom match method" do
+    uri = "https://a.app.co/"
+    client_uri = "https://*.app.co/"
+
+    assert RedirectURI.matches?(uri, client_uri, redirect_uri_match_fun: fn uri, %{host: "*." <> host} = client_uri, _config ->
+      String.ends_with?(uri.host, host) && %{uri | query: nil} == %{client_uri | host: uri.host, authority: uri.authority}
+    end)
   end
 
   test "matches?#true ignores query parameter on comparison" do
-    assert RedirectURI.matches?("https://app.co/?query=hello", "https://app.co/")
+    assert RedirectURI.matches?("https://app.co/?query=hello", "https://app.co/", [])
   end
 
   test "matches?#false" do
-    refute RedirectURI.matches?("https://app.co/?query=hello", "https://app.co")
+    refute RedirectURI.matches?("https://app.co/?query=hello", "https://app.co", [])
   end
 
   test "matches?#false with domains that doesn't start at beginning" do
-    refute RedirectURI.matches?("https://app.co/?query=hello", "https://example.com?app.co=test")
+    refute RedirectURI.matches?("https://app.co/?query=hello", "https://example.com?app.co=test", [])
   end
 
   test "valid_for_authorization?#true" do
@@ -72,7 +81,7 @@ defmodule ExOauth2Provider.RedirectURITest do
   test "valid_for_authorization?#true with custom match method" do
      uri = "https://a.app.co/"
      client_uri = "https://*.app.co/"
- 
+
      config = [
        {
          :redirect_uri_match_fun,
@@ -82,7 +91,7 @@ defmodule ExOauth2Provider.RedirectURITest do
          end
        }
      ]
- 
+
      assert RedirectURI.valid_for_authorization?(uri, client_uri, config)
    end